requirements.txt
streamlit>=1.32.0

# LangChain (modular setup)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-ollama>=0.1.0

# Vector DB
faiss-cpu>=1.7.4

# Ollama client
ollama>=0.1.7

# PDF processing
pymupdf>=1.23.0
pdfplumber>=0.10.3
pypdf>=4.0.0

# OCR (for scanned PDFs)
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# DOCX support
python-docx>=1.1.0

# Utils
requests>=2.31.0